No need to remove callback from map when unregistering a shortcut
The relevant event will no longer be triggered after the shortcut is unregistered. If the same key is registered again, the original callback will be replaced with new one.
Therefore, we don't need to find the registered key from the map. Besides, we don't need to parse the accelerator, too.

An error will be returned if the method failed.